FIX refactor mongo logic
