+++ src/lib/mongoBackend/MongoCommonUpdate.cpp
@@ -51,6 +51,7 @@
 #include "rest/uriParamNames.h"
 
 #include "mongoBackend/connectionOperations.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/safeMongo.h"
 #include "mongoBackend/dbConstants.h"
 #include "mongoBackend/dbFieldEncoding.h"
