+++ src/lib/mongoBackend/MongoGlobal.cpp
@@ -78,7 +78,6 @@ using mongo::BSONObj;
 using mongo::BSONElement;
 using mongo::BSONArrayBuilder;
 using mongo::BSONObjBuilder;
-using mongo::Query;
 using mongo::AssertionException;
 using mongo::BSONArray;
 using mongo::OID;
@@ -277,81 +276,6 @@ void mongoInit
       ensureDateExpirationIndex(tenant);
     }
   }
-}
-
-
-
-#ifdef UNIT_TEST
-
-static DBClientBase* connection = NULL;
-
-
-
-/* ****************************************************************************
-*
-* setMongoConnectionForUnitTest -
-*
-* For unit tests there is only one connection. This connection is stored right here (DBClientBase* connection) and
-* given out using the function getMongoConnection().
-*/
-void setMongoConnectionForUnitTest(DBClientBase* _connection)
-{
-  connection = _connection;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoInitialConnectionGetForUnitTest -
-*
-* This function is meant to be used by unit tests, to get a connection from the pool
-* and then use that connection, setting it with the function 'setMongoConnectionForUnitTest'.
-* This will set the static variable 'connection' in MongoGlobal.cpp and later 'getMongoConnection'
-* returns that variable (getMongoConnection is used by the entire mongo backend).
-*/
-DBClientBase* mongoInitialConnectionGetForUnitTest(void)
-{
-  return mongoPoolConnectionGet();
-}
-#endif
-
-
-
-/* ****************************************************************************
-*
-* getMongoConnection -
-*
-* I would prefer to have per-collection methods, to have a better encapsulation, but
-* the Mongo C++ API doesn't seem to work that way
-*/
-DBClientBase* getMongoConnection(void)
-{
-#ifdef UNIT_TEST
-  return connection;
-#else
-  return mongoPoolConnectionGet();
-#endif
-}
-
-
-
-/* ****************************************************************************
-*
-* releaseMongoConnection - give back mongo connection to connection pool
-*
-* Older versions of this function planned to use a std::auto_ptr<DBClientCursor>* parameter
-* in order to invoke kill() on it for a "safer" connection releasing. However, at the end
-* it seems that kill() will not help at all (see deetails in https://github.com/telefonicaid/fiware-orion/issues/1568)
-* and after some testing we have checked that the current solution is stable.
-*/
-void releaseMongoConnection(DBClientBase* connection)
-{
-#ifdef UNIT_TEST
-  return;
-#else
-  return mongoPoolConnectionRelease(connection);
-#endif  // UNIT_TEST
 }
 
 
@@ -1581,11 +1505,13 @@ bool entitiesQuery
 
   /* Do the query on MongoDB */
   std::auto_ptr<DBClientCursor>  cursor;
-  Query                          query(finalQuery.obj());
+
+  BSONObj query = finalQuery.obj();
+  BSONObj sort;
 
   if (sortOrderList == "")
   {
-    query.sort(BSON(ENT_CREATION_DATE << 1));
+    sort = BSON(ENT_CREATION_DATE << 1);
   }
   else if ((sortOrderList == ORDER_BY_PROXIMITY))
   {
@@ -1619,13 +1545,13 @@ bool entitiesQuery
       sortOrder.append(sortCriteria(sortToken), sortDirection);
     }
 
-    query.sort(sortOrder.obj());
+    sort = sortOrder.obj();
   }
 
   TIME_STAT_MONGO_READ_WAIT_START();
   DBClientBase* connection = getMongoConnection();
 
-  if (!collectionRangedQuery(connection, getEntitiesCollectionName(tenant), query, limit, offset, &cursor, countP, err))
+  if (!collectionRangedQuery(connection, getEntitiesCollectionName(tenant), query, sort, limit, offset, &cursor, countP, err))
   {
     releaseMongoConnection(connection);
     TIME_STAT_MONGO_READ_WAIT_STOP();
@@ -2130,9 +2056,8 @@ bool registrationsQuery
   //           contextRegistration array (e.g. "expiration" is not needed)
   //
   std::auto_ptr<DBClientCursor>  cursor;
-  Query                          query(queryBuilder.obj());
 
-  query.sort(BSON("_id" << 1));
+  BSONObj query = queryBuilder.obj();
 
   LM_T(LmtPagination, ("Offset: %d, Limit: %d, Details: %s", offset, limit, (details == true)? "true" : "false"));
 
@@ -2140,7 +2065,7 @@ bool registrationsQuery
   DBClientBase* connection = getMongoConnection();
   std::string   colName    = getRegistrationsCollectionName(tenant);
 
-  if (!collectionRangedQuery(connection, colName, query, limit, offset, &cursor, countP, err))
+  if (!collectionRangedQuery(connection, colName, query, BSON("_id" << 1), limit, offset, &cursor, countP, err))
   {
     releaseMongoConnection(connection);
     TIME_STAT_MONGO_READ_WAIT_STOP();
