+++ src/lib/mongoBackend/mongoConnectionPool.h
@@ -108,4 +108,31 @@ extern const char* mongoConnectionPoolSemGet(void);
 */
 extern const char* mongoConnectionSemGet(void);
 
+
+
+// Higher level functions (previously in MongoGlobal)
+
+#ifdef UNIT_TEST
+extern void setMongoConnectionForUnitTest(mongo::DBClientBase* _connection);
+#endif
+
+
+/* ****************************************************************************
+*
+* getMongoConnection -
+*
+* I would prefer to have per-collection methods, to have a better encapsulation, but
+* the Mongo C++ API seems not to work that way
+*/
+extern mongo::DBClientBase* getMongoConnection(void);
+
+
+
+/* ****************************************************************************
+*
+* releaseMongoConnection -
+*/
+extern void releaseMongoConnection(mongo::DBClientBase* connection);
+
+
 #endif  // SRC_LIB_MONGOBACKEND_MONGOCONNECTIONPOOL_H_
