+++ src/lib/mongoBackend/mongoQueryTypes.cpp
@@ -35,6 +35,7 @@
 
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/connectionOperations.h"
+#include "mongoConnectionPool.h"
 #include "mongoBackend/dbFieldEncoding.h"
 #include "mongoBackend/safeMongo.h"
 #include "mongoBackend/mongoQueryTypes.h"
