+++ test/unittests/mongoBackend/mongoCreateSubscription_test.cpp
@@ -33,6 +33,7 @@
 #include "logMsg/traceLevels.h"
 #include "common/globals.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoCreateSubscription.h"
 
 #include "unittests/testInit.h"
