+++ test/unittests/mongoBackend/mongoGetSubscriptions_test.cpp
@@ -32,6 +32,7 @@
 #include "logMsg/traceLevels.h"
 #include "common/globals.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoGetSubscriptions.h"
 
 #include "unittests/testInit.h"
