+++ test/unittests/mongoBackend/mongoNotifyContext_test.cpp
@@ -33,6 +33,7 @@
 #include "common/globals.h"
 #include "apiTypesV2/HttpInfo.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoNotifyContext.h"
 #include "ngsi10/NotifyContextRequest.h"
