+++ test/unittests/mongoBackend/mongoQueryContext_filters_test.cpp
@@ -32,6 +32,7 @@
 #include "common/globals.h"
 #include "orionTypes/OrionValueType.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoQueryContext.h"
 #include "ngsi/EntityId.h"
 #include "ngsi10/QueryContextRequest.h"
