+++ test/unittests/mongoBackend/mongoRegisterContext_test.cpp
@@ -33,6 +33,7 @@
 
 #include "common/globals.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoRegisterContext.h"
 #include "ngsi/StatusCode.h"
 #include "ngsi/ContextRegistration.h"
