+++ test/unittests/mongoBackend/mongoUpdateContext_test.cpp
@@ -34,6 +34,7 @@
 #include "common/errorMessages.h"
 #include "orionTypes/OrionValueType.h"
 #include "mongoBackend/MongoGlobal.h"
+#include "mongoBackend/mongoConnectionPool.h"
 #include "mongoBackend/mongoUpdateContext.h"
 #include "mongoBackend/mongoQueryContext.h"
 #include "ngsi/EntityId.h"
